+++ R/cpp11.R
@@ -28,8 +28,8 @@ connect_adf_ <- function(filename, write_protected) {
   .Call(`_adfExplorer2_connect_adf_`, filename, write_protected)
 }
 
-adf_dev_format <- function(connection, name, ffs, intl, dircache) {
-  .Call(`_adfExplorer2_adf_dev_format`, connection, name, ffs, intl, dircache)
+adf_dev_format <- function(connection, name, ffs, intl, dircache, bootable) {
+  .Call(`_adfExplorer2_adf_dev_format`, connection, name, ffs, intl, dircache, bootable)
 }
 
 adf_dev_name <- function(connection, vol_num) {
@@ -100,10 +100,18 @@ adf_entry_to_path <- function(connection, vol_num, sectnum, full) {
   .Call(`_adfExplorer2_adf_entry_to_path`, connection, vol_num, sectnum, full)
 }
 
-adf_dir_list <- function(connection, filename) {
-  .Call(`_adfExplorer2_adf_dir_list`, connection, filename)
+adf_dir_list_ <- function(connection, filename, recursive) {
+  .Call(`_adfExplorer2_adf_dir_list_`, connection, filename, recursive)
 }
 
 adf_change_dir <- function(connection, path) {
   .Call(`_adfExplorer2_adf_change_dir`, connection, path)
+}
+
+adf_get_current_dir <- function(connection) {
+  .Call(`_adfExplorer2_adf_get_current_dir`, connection)
+}
+
+adf_mkdir <- function(connection, path) {
+  .Call(`_adfExplorer2_adf_mkdir`, connection, path)
 }