+++ R/init.R
@@ -1,0 +1,4 @@
+setOldClass("virtual_path")
+setOldClass("adf_device")
+setOldClass("adf_file_con")
+setOldClass("adf_block")