+++ src/file_conn_adf.cpp
@@ -164,7 +164,12 @@ r_string adf_file_con_info(SEXP connection) {
   
   adf_file_con_str *afc = (adf_file_con_str *) con->private_ptr;
   AdfFile *af = afc->adf_file;
-  SEXP parent = getCon(afc->parent_con);
+  sexp parent = getCon(afc->parent_con);
+  
+  writable::strings cls = strings(parent.attr("class"));
+  cls.insert(0, "adf_device");
+  parent.attr("class") = cls;
+
   std::string path = adf_entry_to_path(parent, afc->vol_num,
                                        af->fileHdr->headerKey, TRUE);
   std::string access = "read only";
@@ -175,7 +180,6 @@ r_string adf_file_con_info(SEXP connection) {
 }
 
 void adf_file_con_test_class(SEXP connection) {
-  auto inherits = cpp11::package("base")["inherits"];
-  if (! inherits(connection, "adf_file_con"))
+  if (! Rf_inherits(connection, "adf_file_con"))
     Rf_error("Connection should inherit 'adf_file_con'.");
 }