+++ NEWS.md
@@ -1,4 +1,10 @@
-adfExplorer2 v0.0.9001
+adfExplorer2 v0.0.9002
 -------------
 
-  * Initial commit
+  * Initial commit, which can:
+    * Connect to ADF and ADZ files
+    * Create new virtual floppy disks (including file system)
+    * Connect to files on virtual floppy disks
+    * Obtain device information
+    * List files and directories on virtual device
+    * Create directories on virtual device